+++ core/server/src/binary/handlers/consumer_groups/get_consumer_group_handler.rs
@@ -54,8 +54,8 @@ impl ServerCommandHandler for GetConsumerGroup {
             .with_stream_by_id(&self.stream_id, streams::helpers::get_stream_id());
         shard.permissioner.borrow().get_consumer_group(
             session.get_user_id(),
-            numeric_stream_id as u32,
-            numeric_topic_id as u32,
+            numeric_stream_id,
+            numeric_topic_id,
         )?;
 
         let consumer_group = shard.streams2.with_consumer_group_by_id(
